requirements.txt
networkx==2.8.4
numpy==1.23.5
PyYAML==6.0
PyYAML==6.0.1
scipy==1.10.0
torch==2.1.2+cu118
torch_geometric==2.4.0
torch_scatter==2.1.2+pt21cu118
torch_sparse==0.6.18+pt21cu118
tqdm==4.64.1
